Add files via upload
This python script can extract the text from the .pcap file. Also remeber to save both the python file and pcap file in the same folder. If you want to  change the name of the file open the file in any text editor and change the "shield.pcap" into your file name.